+++ js/lib/helpers.js
@@ -16,3 +16,16 @@ export const formatUpdatesData = updates => {
     }, [])
     .sort((a, b) => a.title - b.title);
 };
+
+
+export const formatLearningCurriculumData = allSubjects => {
+  return allSubjects
+    .reduce((acc, curr) => {
+      const subjectExists = acc.find(subject => subject.title === curr.subject);
+      subjectExists
+        ? subjectExists.data.push(curr)
+        : acc.push({ title: curr.title, data: [curr] });
+      return acc;
+    }, [])
+    .sort((a, b) => a.title - b.title);
+};
